Sepração do ApiService em classes distintas, ajustes no código retirando o que não era necessário.
