+++ RankingVendedores/Pages/Rankings.razor
@@ -4,8 +4,6 @@
 @using Ranking.Aplicacao.Interfaces
 @using Ranking.Aplicacao
 @using Ranking.Dominio.Enums
-@using RankingVendedores.Services
-@inject IApiService ApiService
 @inject ISnackbar Snackbar
 @inject IDialogService DialogService
 @inject RankingViewModel ViewModel
