+++ RankingVendedores/Pages/Vendas.razor
@@ -105,6 +105,7 @@
             <MudTh>Valor</MudTh>
             <MudTh>Data</MudTh>
             <MudTh>Descrição</MudTh>
+            <MudTh></MudTh>
         </HeaderContent>
         <RowTemplate>
             <MudTd>@context.NomeFuncionario</MudTd>
