+++ RankingVendedores/ViewModels/MetaViewModel.cs
@@ -1,16 +1,16 @@
 ﻿using Aplicacao.Utils;
 using MudBlazor;
 using Ranking.Aplicacao.DTOs;
-using Ranking.Aplicacao.Servicos;
 using Ranking.Aplicacao.Validacoes;
-using RankingVendedores.Services;
+using RankingVendedores.Servicos.Interfaces;
 using System.Collections.ObjectModel;
 
 namespace RankingVendedores.ViewModels
 {
     public class MetaViewModel : ViewModelBase
     {
-        private readonly IApiService _apiService;
+        private readonly IMetaApiService _apiService;
+        private readonly IIndicadorApiService _indicadorApiService;
         private readonly ValidadorCriarMeta _validadorCriarMeta = new();
 
         public List<MetaDto> Metas { get; private set; } = new();
@@ -80,7 +80,7 @@ namespace RankingVendedores.ViewModels
 
         public string? MensagemErro { get; private set; }
 
-        public MetaViewModel(IApiService apiService)
+        public MetaViewModel(IMetaApiService apiService)
         {
             _apiService = apiService ?? throw new ArgumentNullException(nameof(apiService));
         }
@@ -133,7 +133,7 @@ namespace RankingVendedores.ViewModels
         {
             try
             {
-                var indicadores = await _apiService.ObterIndicadoresAsync();
+                var indicadores = await _indicadorApiService.ObterIndicadoresAsync();
                 IndicadoresDisponiveis.Clear();
 
                 foreach (var indicador in indicadores.Dados)
